requirements.txt
Flask
line-bot-sdk
gunicorn
apscheduler==3.10.4